+++ src/main/java/randoop/main/GenInputsAbstract.java
@@ -653,14 +653,14 @@ public abstract class GenInputsAbstract extends CommandHandler {
   }
 
   /** Whether to use literals from all classes under test to generate tests. */
-  @Option("Whether to enable Constant Mining to extract constants from SUT")
+  @Option("Whether to enable Constant-TF-IDF for selecting constants as procedure inputs")
   public static boolean constant_tfidf = false;
 
   /**
    * The probability of using a constant value as an input to a method under test. This option is
    * only used when {@code --constant-tfidf} is set to true.
    */
-  @Option("The probability to use Constant Mining")
+  @Option("The probability to use Constant-TF-IDF")
   public static double constant_tfidf_probability = 0.1;
 
   /**
@@ -1038,7 +1038,9 @@ public abstract class GenInputsAbstract extends CommandHandler {
     if (constant_tfidf && (constant_tfidf_probability < 0 || constant_tfidf_probability > 1)) {
       throw new RandoopUsageError(
           "Invalid parameter combination:"
-              + " specified --constant-tfidf and --constant-tfidf-probability is not in [0, 1]");
+              + " specified --constant-tfidf and --constant-tfidf-probability="
+              + constant_tfidf_probability
+              + " is not in [0, 1]");
     }
 
     if (deterministic && ReflectionExecutor.usethreads) {
