+++ src/systemTest/java/randoop/main/RandoopSystemTest.java
@@ -721,22 +721,22 @@ public class RandoopSystemTest {
   }
 
   @Test
-  public void runConstantMiningTest() {
+  public void runConstantTfIdfTest() {
     SystemTestEnvironment testEnvironment =
         systemTestEnvironmentManager.createTestEnvironment("constant-tfidf-test"); // temp directory
     RandoopOptions options = createRandoopOptions(testEnvironment);
     options.setPackageName(null);
-    options.setRegressionBasename("ConstantMiningTest");
-    options.setErrorBasename("ConstantMiningErr");
+    options.setRegressionBasename("ConstantTfIdfTest");
+    options.setErrorBasename("ConstantTfIdfErr");
 
     options.setOption("attempted_limit", "1000");
     options.setOption("generated_limit", "100");
-    options.addTestClass("constantmining.hospital.Doctor");
-    options.addTestClass("constantmining.hospital.Patient");
-    options.addTestClass("constantmining.hospital.AgeConstants");
-    options.addTestClass("constantmining.pharmacy.MedicationConstants");
-    options.addTestClass("constantmining.pharmacy.Pharmacist");
-    options.addTestClass("constantmining.pharmacy.Prescription");
+    options.addTestClass("constanttfidf.hospital.Doctor");
+    options.addTestClass("constanttfidf.hospital.Patient");
+    options.addTestClass("constanttfidf.hospital.AgeConstants");
+    options.addTestClass("constanttfidf.pharmacy.MedicationConstants");
+    options.addTestClass("constanttfidf.pharmacy.Pharmacist");
+    options.addTestClass("constanttfidf.pharmacy.Prescription");
 
     options.setOption("literals-level", "ALL");
     options.setOption("constant-tfidf", "true");
