+++ src/testInput/java/constanttfidf/hospital/AgeConstants.java
@@ -1,4 +1,4 @@
-package constantmining.hospital;
+package constanttfidf.hospital;
 
 // AgeConstants.java
 public class AgeConstants {
